+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Current
 
+* template `left-col`: fixed bug affix
 * fixed `share.mail` in `locales.js`
 
 ## 0.12.6 (2015-06-16)